requirements.txt
# app dependencies

Faker==9.0.0
Flask==3.0.0
Flask-Migrate==4.0.5
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.1.1
Jinja2==3.1.2
python-dotenv==1.0.0
SQLAlchemy==2.0.23